# myordersapi/public/swagger.yaml
openapi: 3.0.3
info:
  title: Order API
  description: Simple Order API with JWT Auth
  version: "1.0.0"

servers:
  - url: http://localhost:8000

tags:
  - name: Auth
    description: Authentication & registration
  - name: Orders
    description: Order CRUD and workflow

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OrderLine:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        productName:
          type: string
        pickedDate:
          type: string
          format: date-time
          nullable: true
        currency:
          type: string

    Task:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        executionDate:
          type: string
          format: date-time
          nullable: true
    Order:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        orderNumber:
          type: integer
        orderDate:
          type: string
          format: date-time
        status:
          type: string
        currency:
          type: string
          readOnly: true
        total:
          type: number
          readOnly: true
        orderLines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLine"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"

security:
  - bearerAuth: []

paths:
  /register:
    post:
      tags: [Auth]
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: test123
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          description: Missing data
        "409":
          description: User already exists

  /api/login_check:
    post:
      tags: [Auth]
      summary: Login user (get JWT token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: test123
      responses:
        "200":
          description: JWT Token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "401":
          description: Invalid credentials

  /api/orders:
    post:
      tags: [Orders]
      summary: Create an order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [name, orderNumber, orderDate, status, currency, orderLines]
              properties:
                name:
                  type: string
                orderNumber:
                  type: integer
                orderDate:
                  type: string
                  format: date-time
                status:
                  type: string
                currency:
                  type: string
                orderLines:
                  type: array
                  items:
                    $ref: "#/components/schemas/OrderLine"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

    get:
      tags: [Orders]
      summary: List all orders for user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

    put:
      tags: [Orders]
      summary: Update order by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

    delete:
      tags: [Orders]
      summary: Delete order by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  /api/orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /api/orders/{id}/tasks:
    post:
      tags: [Orders]
      summary: Link tasks to order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tasks]
              properties:
                tasks:
                  type: array
                  items:
                    $ref: "#/components/schemas/Task"
      responses:
        "200":
          description: Tasks linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /api/orders/{orderId}/tasks/{taskId}:
    put:
      tags: [Orders]
      summary: Update a specific task on an order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                executionDate:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        "200":
          description: Task updated, returns full order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

    delete:
      tags: [Orders]
      summary: Delete a specific task on an order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "404":
          description: Not found